requirements.txt
# 核心RAG框架
llama-index==0.10.12
llama-index-vector-stores-faiss==0.1.2
llama-index-embeddings-huggingface==0.2.0

# 向量数据库
faiss-cpu==1.7.4

# 嵌入模型
sentence-transformers>=2.6.1
transformers>=4.36.2
torch>=2.1.1

# LLM API客户端
openai==1.6.1
requests==2.31.0

# 数据处理
pandas==2.1.4
numpy==1.24.3

# 评估工具
scikit-learn==1.3.2

# 其他工具
python-dotenv==1.0.0
tqdm==4.66.1
